fixed crash on build dir not existing in clean.py

+++ buildscripts/clear.py
@@ -4,12 +4,13 @@ import sys
 
 def clean():
     try:
-        shutil.rmtree("build")
+        if os.path.exists("build"):
+            shutil.rmtree("build")
         os.mkdir("build")
     except PermissionError:
-        print("Error: Permission denied! Continueing...", file=sys.stderr)
+        print("Error: Permission denied! Continuing...", file=sys.stderr)
     except FileExistsError:
-        print("Error: File not found! Continueing...", file=sys.stderr)
+        print("Error: File not found! Continuing...", file=sys.stderr)
 
 if __name__ == "__main__":
     clean()
